refactor: Gradle 멀티모듈 일반사용자, 관리자 분리
