+++ README.md
@@ -2,29 +2,4 @@
 
 누구나 자신의 상점을 개설할 수 있는 쇼핑몰
 
-## 기술 스택
-
-**Client:** Bootstrap, HTML, CSS, JavaScript
-
-**Server**
- - Java 11
- - Spring Boot 2.5.6
- - Spring Data JPA 
- - Thymeleaf
-
-**Build Tool**
- - Gradle
-
-**DataBase** 
- - MySQL
-
-**Cooperation Tool**
- - Google Docs
- - Git
-
-**CI/CD**
- - Jenkins
-
-**Infra**
- - AWS EC2
- - AWS RDS
+## [1. 기술 스택](https://github.com/mk2e/mk2e-docs/tree/master/01_tech_stack#readme)
