+++ module-admin/README.md
@@ -1,0 +1,5 @@
+# Gradle 멀티모듈 admin
+
+## 관리자가 실제 이용하게 되는 웹애플리케이션입니다.
+
+> 관리자 페이지와 일반사용자 페이지의 의존성을 낮추기 위해 멀티모듈을 이용하였습니다.
