+++ module-web/src/main/java/web/HomeController.java
@@ -1,4 +1,4 @@
-package org.pnm.marketeveryone;
+package web;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
