+++ module-web/src/main/resources/static/index.html
