+++ module-web/src/main/resources/templates/layout/base.html
@@ -13,6 +13,7 @@
 </div>
 <footer th:replace="layout/footer :: common_footer"></footer>
 </body>
+
 <!-- Bootstrap core JavaScript -->
 <script th:src="@{/vendor/jquery/jquery.min.js}"></script>
 <script th:src="@{/vendor/bootstrap/js/bootstrap.bundle.min.js}"></script>
