+++ module-web/src/main/resources/templates/layout/sidebar.html
@@ -1,10 +1,5 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
 <div th:fragment="common_sidebar" class="col-lg-3">
     <h1 class="my-4">카테고리</h1>
     <div class="list-group">
@@ -13,5 +8,4 @@
         <a href="#" class="list-group-item">Category 3</a>
     </div>
 </div>
-</body>
 </html>
