+++ settings.gradle
@@ -1,1 +1,4 @@
 rootProject.name = 'market-everyone'
+include 'module-web'
+include 'module-admin'
+
