--- src/main/resources/templates/home.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org" th:replace="layout/base :: layout (~{::title}, ~{::section})">
-<head>
-    <meta charset="UTF-8">
-    <title>마켓에브리원-메인</title>
-</head>
-<section>
-    홈입니다 홈
-</section>
-</html>
