--- src/test/java/org/pnm/marketeveryone/MarketEveryoneApplicationTests.java
@@ -1,13 +1,0 @@
-package org.pnm.marketeveryone;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MarketEveryoneApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
